requirements.txt
aiogram==3.4.*
langchain==0.1.*
langchain-openai==0.1.*
tavily-python==0.3.*
openai==1.12.*
sqlalchemy==2.0.*
python-dotenv==1.0.*
alembic==1.13.*
loguru==0.7.*
fastapi==0.109.*
uvicorn[standard]==0.27.*